+++ lib/spear/persistent_subscription.ex
@@ -4,6 +4,34 @@ defmodule Spear.PersistentSubscription do
   """
 
   @typedoc """
+  The action the EventStoreDB should take for an event's nack
+
+  * `:park` - stops the EventStoreDB from re-sending the event and appends a
+    reference to the event to the persistent subscription's parked events
+    stream. These events can be retried later by clicking the "Replay
+    Parked Messages" button in the EventStoreDB dashboard's persistent
+    subscriptions section for each stream+group combination. The EventStoreDB
+    parking system is conceptually similar to dead letter queues.
+  * `:retry` - retries the event up to `:max_retry_count` tries. This option is
+    a reasonable default so that if a consumer hits a transient error condition
+    such as a network timeout, it will retry the event before giving up and
+    parking. Note that once a consumer has retried an event more than
+    `:max_retry_count` times, the event is parked, even if the `:retry` action
+    is given to `Spear.nack/4`.
+  * `:skip` - skips the event without moving it to the parked events stream.
+    Practically this is no different than simply `Spear.ack/3`ing the event(s)
+    and performing a no-op in the consumer. Skipping may be a good option for
+    dealing with poison messages: malformed or otherwise completely
+    unhandleable events.
+  * `:stop` - stops the EventStoreDB from re-sending the event. The event is
+    not parked or retried. It is unclear how this differs from the `:skip`
+    action. This function does not stop the persistent subscription:
+    use `Spear.cancel_subscription/3` to shut down the connection.
+  """
+  @typedoc since: "0.6.0"
+  @type nack_action :: :unknown | :park | :retry | :skip | :stop
+
+  @typedoc """
   A persistent subscription.
 
   These are generally returned from `Spear.list_persistent_subscriptions/2`.
@@ -74,4 +102,11 @@ defmodule Spear.PersistentSubscription do
 
   defp to_atom(string) when is_binary(string), do: String.to_atom(string)
   defp to_atom(_), do: nil
+
+  @doc false
+  def map_nack_action(:park), do: :Park
+  def map_nack_action(:retry), do: :Retry
+  def map_nack_action(:skip), do: :Skip
+  def map_nack_action(:stop), do: :Stop
+  def map_nack_action(_), do: :Unknown
 end